requirements.txt
python-Levenshtein==0.12.0
coverage
